[1] Eliminated single warning by fixing print format. [2] Functional and sanity tests PASS.

+++ Hazer/app/gpstool/main.c
@@ -260,7 +260,7 @@ int main(int argc, char * argv[])
     /*
      * Network variables.
      */
-    uint64_t network_total = 0;
+    ssize_t network_total = 0;
     /*
      * Keepalive variables.
      */
